Move api init to own class
