+++ src/server.js
@@ -15,10 +15,9 @@ const { initDB } = require('./db');
 const { initLogger, getLogger } = require('./utils/logger');
 const EmitterHelper = require('./utils/emitterHelper');
 const schema = require('./schema');
-const syncRouter = require('./route/sync');
-const apiRouter = require('./route/api');
 const { startSync } = require('./sync');
 const { getInstance } = require('./qclient');
+const { initApiServer } = require('./route');
 const Wallet = require('./api/wallet');
 
 const walletEncryptedMessage = 'Your wallet is encrypted. Please use a non-encrypted wallet for the server.';
@@ -234,34 +233,6 @@ function startQtumProcess(reindex) {
   }
 }
 
-// Create Restify server and apply routes
-async function startAPI() {
-  server = restify.createServer({ title: 'Bodhi API' });
-  const cors = corsMiddleware({ origins: ['*'] });
-  server.pre(cors.preflight);
-  server.use(cors.actual);
-  server.use(restify.plugins.bodyParser({ mapParams: true }));
-  server.use(restify.plugins.queryParser());
-  server.on('after', (req, res, route, err) => {
-    if (route) {
-      getLogger().debug(`${route.methods[0]} ${route.spec.path} ${res.statusCode}`);
-    } else {
-      getLogger().error(`${err.message}`);
-    }
-  });
-
-  syncRouter.applyRoutes(server);
-  apiRouter.applyRoutes(server);
-
-  server.listen(Config.PORT, () => {
-    SubscriptionServer.create(
-      { execute, subscribe, schema },
-      { server, path: '/subscriptions' },
-    );
-    getLogger().info(`Bodhi API is running at http://${Config.HOSTNAME}:${Config.PORT}.`);
-  });
-}
-
 // Ensure API is running before loading UI
 async function checkApiInit() {
   try {
@@ -282,7 +253,7 @@ async function checkApiInit() {
 
 function startServices() {
   startSync();
-  startAPI();
+  initApiServer();
 
   checkApiInterval = setInterval(checkApiInit, 500);
 }
